core: ardupilot-manager: Move Singleton class to `commonwealth`
