+++ core/libs/commonwealth/commonwealth/utils/Singleton.py
